Moving analysis code to analysis/__init__ for easier importing
